:truck: Rename DashboardComponent to OldDashboardComponent to make room for the new component
