+++ src/app/onion/old-dashboard/components/changelog-item/changelog-item.component.scss
