+++ src/app/onion/old-dashboard/components/dashboard-item/dashboard-item.component.scss
