+++ src/app/onion/old-dashboard/components/identification-pill/identification-pill.component.ts
