+++ src/app/onion/old-dashboard/components/side-panel-content/learning-object/learning-object.component.html
