+++ src/app/onion/old-dashboard/components/side-panel-content/side-panel-content.component.html
