+++ src/app/onion/old-dashboard/components/side-panel/panel.directive.ts
