+++ src/app/onion/old-dashboard/components/side-panel/side-panel-viewer/side-panel-viewer.component.animations.ts
