+++ src/app/onion/old-dashboard/components/side-panel/side-panel.module.ts
