+++ src/app/onion/old-dashboard/old-dashboard.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { DashboardComponent } from './dashboard.component';
+import { OldDashboardComponent } from './old-dashboard.component';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { RouterTestingModule } from '@angular/router/testing';
 import { TooltipModule } from 'app/shared/tooltips/tip.module';
@@ -15,14 +15,14 @@ import { ContextMenuModule } from 'app/shared/contextmenu/contextmenu.module';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { User } from '@entity';
 
-describe('DashboardComponent', () => {
-  let component: DashboardComponent;
-  let fixture: ComponentFixture<DashboardComponent>;
+describe('OldDashboardComponent', () => {
+  let component: OldDashboardComponent;
+  let fixture: ComponentFixture<OldDashboardComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
-      declarations: [ DashboardComponent ],
+      declarations: [ OldDashboardComponent ],
       imports: [
         TooltipModule,
         RouterTestingModule,
@@ -43,7 +43,7 @@ describe('DashboardComponent', () => {
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(DashboardComponent);
+    fixture = TestBed.createComponent(OldDashboardComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
